+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 ## Changelog
 
+### Newest version: 0.1.3
+
+* add inner class check
+* add private constructor check
+
+### Newest version: 0.1.2
+
+* add Set<String>
+* fixed boxed primitive null
+
 ### Newest version: 0.1.1
 
 * Save model as SharedPreferences(private static complex field not support)